requirements.txt
pip~=24.0
openai
langchain
streamlit
langchain_openai
langchain_core
langchain_community
python-dotenv~=1.0.1
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
faiss-cpu
groq
cassio
beautifulsoup4
langchain-groq
wikipedia
arxiv
langchainhub
sentence_transformers
PyPDF2
langchain-objectbox
requests
chromadb